update server with float timeout

+++ import/http/receiver/src/config.rs
@@ -8,7 +8,7 @@ const DEFAULT_PATH: &str = "/etc/import-http/config.toml";
 pub struct Config {
     pub import_address: SocketAddrV4,
     pub export_address: SocketAddrV4,
-    pub timeout: u64,
+    pub timeout: f64,
 }
 
 impl Config {
